Improve test suite class names
