+++ testlib/src/test/java/io/spine/testing/TestingSpec.java
@@ -1,12 +1,6 @@
 /*
  * Copyright 2022, TeamDev. All rights reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
  * disclaimer.
@@ -26,19 +20,43 @@
 
 package io.spine.testing;
 
+import com.google.common.testing.NullPointerTester;
+import com.google.protobuf.FieldMask;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static com.google.common.truth.Truth.assertThat;
+import java.util.concurrent.atomic.AtomicInteger;
 
-class StubMessageTest {
+import static com.google.common.truth.Truth.assertThat;
+import static io.spine.testing.Testing.repeat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+@DisplayName("`Testing` utility class should")
+class TestingSpec extends UtilityClassTest<Testing> {
+
+    TestingSpec() {
+        super(Testing.class);
+    }
+
+    @Override
+    protected void configure(NullPointerTester tester) {
+        tester.setDefault(FieldMask.class, FieldMask.getDefaultInstance());
+    }
 
     @Test
-    @DisplayName("can have derived classes with default constructor")
-    void derivedClass() {
-         var msg = new StubMessage() {
-         };
+    @DisplayName("repeat an action a number of times")
+    void repeating() {
+        var expected = TestValues.random(10);
+        var counter = new AtomicInteger(0);
+        repeat(expected, counter::incrementAndGet);
 
-         assertThat(msg).isNotNull();
+        assertThat(counter.get())
+                .isEqualTo(expected);
+    }
+
+    @Test
+    @DisplayName("provide method `halt()` for failing methods that should never be called")
+    void haltMethod() {
+        assertThrows(AssertionError.class, Testing::halt);
     }
 }
